+++ src/runtime/pyobject.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Dynamic;
+using System.Linq;
 using System.Linq.Expressions;
 
 namespace Python.Runtime
@@ -32,6 +33,8 @@ namespace Python.Runtime
         private bool disposed = false;
         private bool _finalized = false;
 
+        internal BorrowedReference Reference => new BorrowedReference(obj);
+
         /// <summary>
         /// PyObject Constructor
         /// </summary>
@@ -43,7 +46,24 @@ namespace Python.Runtime
         /// </remarks>
         public PyObject(IntPtr ptr)
         {
+            if (ptr == IntPtr.Zero) throw new ArgumentNullException(nameof(ptr));
+
             obj = ptr;
+#if TRACE_ALLOC
+            Traceback = new StackTrace(1);
+#endif
+        }
+
+        /// <summary>
+        /// Creates new <see cref="PyObject"/> pointing to the same object as
+        /// the <paramref name="reference"/>. Increments refcount, allowing <see cref="PyObject"/>
+        /// to have ownership over its own reference.
+        /// </summary>
+        internal PyObject(BorrowedReference reference)
+        {
+            if (reference.IsNull) throw new ArgumentNullException(nameof(reference));
+
+            obj = Runtime.SelfIncRef(reference.DangerousGetAddress());
 #if TRACE_ALLOC
             Traceback = new StackTrace(1);
 #endif
@@ -51,7 +71,7 @@ namespace Python.Runtime
 
         // Protected default constructor to allow subclasses to manage
         // initialization in different ways as appropriate.
-
+        [Obsolete("Please, always use PyObject(*Reference)")]
         protected PyObject()
         {
 #if TRACE_ALLOC
@@ -209,6 +229,8 @@ namespace Python.Runtime
         /// </remarks>
         public bool TypeCheck(PyObject typeOrClass)
         {
+            if (typeOrClass == null) throw new ArgumentNullException(nameof(typeOrClass));
+
             return Runtime.PyObject_TypeCheck(obj, typeOrClass.obj);
         }
 
@@ -221,6 +243,8 @@ namespace Python.Runtime
         /// </remarks>
         public bool HasAttr(string name)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+
             return Runtime.PyObject_HasAttrString(obj, name) != 0;
         }
 
@@ -234,6 +258,8 @@ namespace Python.Runtime
         /// </remarks>
         public bool HasAttr(PyObject name)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+
             return Runtime.PyObject_HasAttr(obj, name.obj) != 0;
         }
 
@@ -247,6 +273,8 @@ namespace Python.Runtime
         /// </remarks>
         public PyObject GetAttr(string name)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+
             IntPtr op = Runtime.PyObject_GetAttrString(obj, name);
             if (op == IntPtr.Zero)
             {
@@ -257,7 +285,7 @@ namespace Python.Runtime
 
 
         /// <summary>
-        /// GetAttr Method
+        /// GetAttr Method. Returns fallback value if getting attribute fails for any reason.
         /// </summary>
         /// <remarks>
         /// Returns the named attribute of the Python object, or the given
@@ -265,6 +293,8 @@ namespace Python.Runtime
         /// </remarks>
         public PyObject GetAttr(string name, PyObject _default)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+
             IntPtr op = Runtime.PyObject_GetAttrString(obj, name);
             if (op == IntPtr.Zero)
             {
@@ -285,6 +315,8 @@ namespace Python.Runtime
         /// </remarks>
         public PyObject GetAttr(PyObject name)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+
             IntPtr op = Runtime.PyObject_GetAttr(obj, name.obj);
             if (op == IntPtr.Zero)
             {
@@ -304,6 +336,8 @@ namespace Python.Runtime
         /// </remarks>
         public PyObject GetAttr(PyObject name, PyObject _default)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+
             IntPtr op = Runtime.PyObject_GetAttr(obj, name.obj);
             if (op == IntPtr.Zero)
             {
@@ -323,6 +357,9 @@ namespace Python.Runtime
         /// </remarks>
         public void SetAttr(string name, PyObject value)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (value == null) throw new ArgumentNullException(nameof(value));
+
             int r = Runtime.PyObject_SetAttrString(obj, name, value.obj);
             if (r < 0)
             {
@@ -341,6 +378,9 @@ namespace Python.Runtime
         /// </remarks>
         public void SetAttr(PyObject name, PyObject value)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (value == null) throw new ArgumentNullException(nameof(value));
+
             int r = Runtime.PyObject_SetAttr(obj, name.obj, value.obj);
             if (r < 0)
             {
@@ -358,6 +398,8 @@ namespace Python.Runtime
         /// </remarks>
         public void DelAttr(string name)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+
             int r = Runtime.PyObject_SetAttrString(obj, name, IntPtr.Zero);
             if (r < 0)
             {
@@ -376,6 +418,8 @@ namespace Python.Runtime
         /// </remarks>
         public void DelAttr(PyObject name)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+
             int r = Runtime.PyObject_SetAttr(obj, name.obj, IntPtr.Zero);
             if (r < 0)
             {
@@ -394,6 +438,8 @@ namespace Python.Runtime
         /// </remarks>
         public virtual PyObject GetItem(PyObject key)
         {
+            if (key == null) throw new ArgumentNullException(nameof(key));
+
             IntPtr op = Runtime.PyObject_GetItem(obj, key.obj);
             if (op == IntPtr.Zero)
             {
@@ -413,6 +459,8 @@ namespace Python.Runtime
         /// </remarks>
         public virtual PyObject GetItem(string key)
         {
+            if (key == null) throw new ArgumentNullException(nameof(key));
+
             using (var pyKey = new PyString(key))
             {
                 return GetItem(pyKey);
@@ -447,6 +495,9 @@ namespace Python.Runtime
         /// </remarks>
         public virtual void SetItem(PyObject key, PyObject value)
         {
+            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (value == null) throw new ArgumentNullException(nameof(value));
+
             int r = Runtime.PyObject_SetItem(obj, key.obj, value.obj);
             if (r < 0)
             {
@@ -465,6 +516,9 @@ namespace Python.Runtime
         /// </remarks>
         public virtual void SetItem(string key, PyObject value)
         {
+            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (value == null) throw new ArgumentNullException(nameof(value));
+
             using (var pyKey = new PyString(key))
             {
                 SetItem(pyKey, value);
@@ -482,6 +536,8 @@ namespace Python.Runtime
         /// </remarks>
         public virtual void SetItem(int index, PyObject value)
         {
+            if (value == null) throw new ArgumentNullException(nameof(value));
+
             using (var pyindex = new PyInt(index))
             {
                 SetItem(pyindex, value);
@@ -499,6 +555,8 @@ namespace Python.Runtime
         /// </remarks>
         public virtual void DelItem(PyObject key)
         {
+            if (key == null) throw new ArgumentNullException(nameof(key));
+
             int r = Runtime.PyObject_DelItem(obj, key.obj);
             if (r < 0)
             {
@@ -517,6 +575,8 @@ namespace Python.Runtime
         /// </remarks>
         public virtual void DelItem(string key)
         {
+            if (key == null) throw new ArgumentNullException(nameof(key));
+
             using (var pyKey = new PyString(key))
             {
                 DelItem(pyKey);
@@ -639,10 +699,13 @@ namespace Python.Runtime
         /// </summary>
         /// <remarks>
         /// Invoke the callable object with the given arguments, passed as a
-        /// PyObject[]. A PythonException is raised if the invokation fails.
+        /// PyObject[]. A PythonException is raised if the invocation fails.
         /// </remarks>
         public PyObject Invoke(params PyObject[] args)
         {
+            if (args == null) throw new ArgumentNullException(nameof(args));
+            if (args.Contains(null)) throw new ArgumentNullException();
+
             var t = new PyTuple(args);
             IntPtr r = Runtime.PyObject_Call(obj, t.obj, IntPtr.Zero);
             t.Dispose();
@@ -659,10 +722,12 @@ namespace Python.Runtime
         /// </summary>
         /// <remarks>
         /// Invoke the callable object with the given arguments, passed as a
-        /// Python tuple. A PythonException is raised if the invokation fails.
+        /// Python tuple. A PythonException is raised if the invocation fails.
         /// </remarks>
         public PyObject Invoke(PyTuple args)
         {
+            if (args == null) throw new ArgumentNullException(nameof(args));
+
             IntPtr r = Runtime.PyObject_Call(obj, args.obj, IntPtr.Zero);
             if (r == IntPtr.Zero)
             {
@@ -677,12 +742,15 @@ namespace Python.Runtime
         /// </summary>
         /// <remarks>
         /// Invoke the callable object with the given positional and keyword
-        /// arguments. A PythonException is raised if the invokation fails.
+        /// arguments. A PythonException is raised if the invocation fails.
         /// </remarks>
         public PyObject Invoke(PyObject[] args, PyDict kw)
         {
+            if (args == null) throw new ArgumentNullException(nameof(args));
+            if (args.Contains(null)) throw new ArgumentNullException();
+
             var t = new PyTuple(args);
-            IntPtr r = Runtime.PyObject_Call(obj, t.obj, kw != null ? kw.obj : IntPtr.Zero);
+            IntPtr r = Runtime.PyObject_Call(obj, t.obj, kw?.obj ?? IntPtr.Zero);
             t.Dispose();
             if (r == IntPtr.Zero)
             {
@@ -697,11 +765,13 @@ namespace Python.Runtime
         /// </summary>
         /// <remarks>
         /// Invoke the callable object with the given positional and keyword
-        /// arguments. A PythonException is raised if the invokation fails.
+        /// arguments. A PythonException is raised if the invocation fails.
         /// </remarks>
         public PyObject Invoke(PyTuple args, PyDict kw)
         {
-            IntPtr r = Runtime.PyObject_Call(obj, args.obj, kw != null ? kw.obj : IntPtr.Zero);
+            if (args == null) throw new ArgumentNullException(nameof(args));
+
+            IntPtr r = Runtime.PyObject_Call(obj, args.obj, kw?.obj ?? IntPtr.Zero);
             if (r == IntPtr.Zero)
             {
                 throw new PythonException();
@@ -715,10 +785,14 @@ namespace Python.Runtime
         /// </summary>
         /// <remarks>
         /// Invoke the named method of the object with the given arguments.
-        /// A PythonException is raised if the invokation is unsuccessful.
+        /// A PythonException is raised if the invocation is unsuccessful.
         /// </remarks>
         public PyObject InvokeMethod(string name, params PyObject[] args)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (args == null) throw new ArgumentNullException(nameof(args));
+            if (args.Contains(null)) throw new ArgumentNullException();
+
             PyObject method = GetAttr(name);
             PyObject result = method.Invoke(args);
             method.Dispose();
@@ -731,10 +805,51 @@ namespace Python.Runtime
         /// </summary>
         /// <remarks>
         /// Invoke the named method of the object with the given arguments.
-        /// A PythonException is raised if the invokation is unsuccessful.
+        /// A PythonException is raised if the invocation is unsuccessful.
         /// </remarks>
         public PyObject InvokeMethod(string name, PyTuple args)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (args == null) throw new ArgumentNullException(nameof(args));
+
+            PyObject method = GetAttr(name);
+            PyObject result = method.Invoke(args);
+            method.Dispose();
+            return result;
+        }
+
+        /// <summary>
+        /// InvokeMethod Method
+        /// </summary>
+        /// <remarks>
+        /// Invoke the named method of the object with the given arguments.
+        /// A PythonException is raised if the invocation is unsuccessful.
+        /// </remarks>
+        public PyObject InvokeMethod(PyObject name, params PyObject[] args)
+        {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (args == null) throw new ArgumentNullException(nameof(args));
+            if (args.Contains(null)) throw new ArgumentNullException();
+
+            PyObject method = GetAttr(name);
+            PyObject result = method.Invoke(args);
+            method.Dispose();
+            return result;
+        }
+
+
+        /// <summary>
+        /// InvokeMethod Method
+        /// </summary>
+        /// <remarks>
+        /// Invoke the named method of the object with the given arguments.
+        /// A PythonException is raised if the invocation is unsuccessful.
+        /// </remarks>
+        public PyObject InvokeMethod(PyObject name, PyTuple args)
+        {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (args == null) throw new ArgumentNullException(nameof(args));
+
             PyObject method = GetAttr(name);
             PyObject result = method.Invoke(args);
             method.Dispose();
@@ -748,10 +863,14 @@ namespace Python.Runtime
         /// <remarks>
         /// Invoke the named method of the object with the given arguments
         /// and keyword arguments. Keyword args are passed as a PyDict object.
-        /// A PythonException is raised if the invokation is unsuccessful.
+        /// A PythonException is raised if the invocation is unsuccessful.
         /// </remarks>
         public PyObject InvokeMethod(string name, PyObject[] args, PyDict kw)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (args == null) throw new ArgumentNullException(nameof(args));
+            if (args.Contains(null)) throw new ArgumentNullException();
+
             PyObject method = GetAttr(name);
             PyObject result = method.Invoke(args, kw);
             method.Dispose();
@@ -765,10 +884,13 @@ namespace Python.Runtime
         /// <remarks>
         /// Invoke the named method of the object with the given arguments
         /// and keyword arguments. Keyword args are passed as a PyDict object.
-        /// A PythonException is raised if the invokation is unsuccessful.
+        /// A PythonException is raised if the invocation is unsuccessful.
         /// </remarks>
         public PyObject InvokeMethod(string name, PyTuple args, PyDict kw)
         {
+            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (args == null) throw new ArgumentNullException(nameof(args));
+
             PyObject method = GetAttr(name);
             PyObject result = method.Invoke(args, kw);
             method.Dispose();
@@ -785,6 +907,8 @@ namespace Python.Runtime
         /// </remarks>
         public bool IsInstance(PyObject typeOrClass)
         {
+            if (typeOrClass == null) throw new ArgumentNullException(nameof(typeOrClass));
+
             int r = Runtime.PyObject_IsInstance(obj, typeOrClass.obj);
             if (r < 0)
             {
@@ -804,6 +928,8 @@ namespace Python.Runtime
         /// </remarks>
         public bool IsSubclass(PyObject typeOrClass)
         {
+            if (typeOrClass == null) throw new ArgumentNullException(nameof(typeOrClass));
+
             int r = Runtime.PyObject_IsSubclass(obj, typeOrClass.obj);
             if (r < 0)
             {